feat(2756): address review feedback.
